Remove web specific method from Repository interface
Exception handling should ultimately move to ExceptionMapper implementations
